- Modified integration code with .NET reflection metadata to be more efficient and avoid potentially wrong resolution - Added MemberTest - Moved Constants to root namespace

+++ Fasterflect/FasterflectTest/SampleModel/People/Employee.cs
@@ -26,7 +26,7 @@
             Subordinates = subordinates;
         }
 
-        void ISwimmable.Swim( int meters )
+        void ISwimmable.Swim( double meters )
         {
             metersTravelled += meters;
         }
